Fetch state in App and pass it down to Pong as props
- Nested components under <Stage> cannot access the store
