+++ src/index.js
@@ -1,16 +1,15 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import "./index.css";
-import App from "./App";
-import CounterApp from "./CounterApp";
-import * as serviceWorker from "./serviceWorker";
-import store from "./app/store";
-
 import { Provider } from "react-redux";
+
+import "./index.css";
+import * as serviceWorker from "./serviceWorker";
+import App from "./App";
+import store from "./app/store";
 
 ReactDOM.render(
   <Provider store={store}>
-    <CounterApp />
+    <App />
   </Provider>,
   document.getElementById("root")
 );
